refactor: 할인 관련 로직을 유틸로 분리하기
