+++ src/refactoring/components/CartPage.tsx
@@ -1,5 +1,6 @@
 import { CartItem, Product } from '../../types.ts';
 import { useCart } from '../hooks';
+import { getAppliedDiscount, getMaxDiscount } from '../utils/discountUtils.ts';
 import { useProductContext } from '../contexts/productContext.tsx';
 import { useCouponContext } from '../contexts/couponContext.tsx';
 
@@ -17,10 +18,6 @@ export const CartPage = () => {
     selectedCoupon,
   } = useCart();
 
-  const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
-    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-  };
-
   const getRemainingStock = (product: Product) => {
     const cartItem = cart.find((item) => item.product.id === product.id);
     return product.stock - (cartItem?.quantity || 0);
@@ -28,18 +25,6 @@ export const CartPage = () => {
 
   const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } =
     calculateTotal();
-
-  const getAppliedDiscount = (item: CartItem) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-      if (quantity >= discount.quantity) {
-        appliedDiscount = Math.max(appliedDiscount, discount.rate);
-      }
-    }
-    return appliedDiscount;
-  };
 
   return (
     <div className="container mx-auto p-4">
